requirements.txt
certifi==2021.10.8
Pillow==8.4.0
pycryptodomex==3.11.0
pydub==0.25.1
pyzipper==0.3.5
tqdm==4.62.3
